modify layer2 chainid to uint64 type

+++ chainbridge-core/blockstore/blockstore.go
@@ -30,7 +30,7 @@ var (
 	ErrNotFound = errors.New("key not found")
 )
 
-func StoreBlock(db KeyValueWriter, block *big.Int, chainID uint8) error {
+func StoreBlock(db KeyValueWriter, block *big.Int, chainID uint64) error {
 	key := bytes.Buffer{}
 	keyS := fmt.Sprintf("chain:%s:block", string(chainID))
 	key.WriteString(keyS)
@@ -41,7 +41,7 @@ func StoreBlock(db KeyValueWriter, block *big.Int, chainID uint8) error {
 	return nil
 }
 
-func GetLastStoredBlock(db KeyValueReader, chainID uint8) (*big.Int, error) {
+func GetLastStoredBlock(db KeyValueReader, chainID uint64) (*big.Int, error) {
 	key := bytes.Buffer{}
 	keyS := fmt.Sprintf("chain:%s:block", string(chainID))
 	key.WriteString(keyS)
